+++ projects/net.splitcells.gel.core/src/main/java/net/splitcells/gel/rating/rater/HasMinimalSize.java
@@ -20,6 +20,7 @@ import net.splitcells.dem.data.set.list.List;
 import net.splitcells.dem.lang.Xml;
 import net.splitcells.dem.lang.dom.Domable;
 import net.splitcells.dem.object.Discoverable;
+import net.splitcells.dem.utils.CommonFunctions;
 import net.splitcells.gel.constraint.Constraint;
 import net.splitcells.gel.constraint.GroupId;
 import net.splitcells.gel.data.table.Line;
@@ -142,6 +143,11 @@ public class HasMinimalSize implements Rater {
             return this.minimalSize == ((HasMinimalSize) arg).minimalSize;
         }
         return false;
+    }
+
+    @Override
+    public int hashCode() {
+        return CommonFunctions.hashCode(minimalSize);
     }
 
     @Override